requirements.txt
2captcha-python==1.0.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
python-decouple==3.4
requests==2.25.1
selenium==3.141.0
urllib3==1.26.6
Pillow==8.3.1